requirements.txt
Django>=5.1
psycopg2>=2.9
django_rest_framework
